Reordered Parser.py order to be able to catch wrong json packets; fixed simulate_sensors and send_json
simulate_sensors use send_json's request function, and it was implying that server will answer a json string, which is not true by now, so throwed json decode errors; now is changed to just print the response, what ever it is.

Parser.py has JSON_parser which responds to the sensor values sent to server, it is important to avoid trying to compare with limits and do other operations over a json that doesn't have the expected format, so first operation is sending data to DB referencing the jdata request variable, if it has the expected format will continue, ValueError or even TypeError will be thrown.