Almost complete the stock graph documentation and migration to command like instructions instead of a big json object
The documentation is C based comment to use doxygen
It's still need to be implemented the logic to add one dataset per sensor type.
The logic should be create one object from type DataSet for each sensor type.
Then we create a list, append all the DataSet objects to it and set that list as a parameter value for the property DataSets of the chart (greenhouse_chart in our case)